refactor: 🔄 画像のディレクトリ構成とパスを更新
